+++ composer.json
@@ -5,10 +5,13 @@
   "authors": [
     {
       "name": "Matt Miller"
+    },
+    {
+      "name": "Marc Bellêtre"
     }
   ],
   "require": {
-    "composer/installers": "^1.0|^2.0"
+    "php": ">=7.1.0"
   },
   "autoload": {
     "psr-4": {
@@ -18,8 +21,8 @@
   "minimum-stability": "dev",
   "prefer-stable": true,
   "config": {
-    "allow-plugins": {
-      "composer/installers": true
-    }
+    "optimize-autoloader": true,
+    "preferred-install": "dist",
+    "sort-packages": true
   }
 }
