+++ index.php
@@ -7,29 +7,35 @@
  * Author URI: www.millermedia.io
  */
 
-use ProtectTheChildren\Admin;
-use ProtectTheChildren\Helpers;
+namespace ProtectTheChildren;
 
-require 'vendor/autoload.php';
+add_action('plugins_loaded', function () {
+    if (file_exists($composer = __DIR__.'/vendor/autoload.php')) {
+        require_once $composer;
+    }
 
-if (! defined('PROTECT_THE_CHILDREN_PLUGIN_VERSION')) {
-    define('PROTECT_THE_CHILDREN_PLUGIN_VERSION', '1.5.0');
-}
+    if (! defined('PROTECT_THE_CHILDREN_PLUGIN_VERSION')) {
+        define('PROTECT_THE_CHILDREN_PLUGIN_VERSION', '1.5.0');
+    }
 
-if (version_compare(PHP_VERSION, '5.6', '<')) {
-    add_action('admin_notices', function () {
-        echo '<div class="error"><p>'.__('Protect the Children requires PHP 5.6 and greater to function properly. Please upgrade PHP or deactivate Protect the Children.', 'protect-the-children').'</p></div>';
-    });
+    if (version_compare(PHP_VERSION, '5.6', '<')) {
+        add_action('admin_notices', function () {
+            echo '<div class="error"><p>'.__('Protect the Children requires PHP 5.6 and greater to function properly. Please upgrade PHP or deactivate Protect the Children.', 'protect-the-children').'</p></div>';
+        });
 
-    return;
-}
+        return;
+    }
 
-define('PTC_PLUGIN_PATH', plugin_dir_path(__FILE__));
-define('PTC_PLUGIN_URL', plugin_dir_url(__FILE__));
+    define('PTC_PLUGIN_PATH', plugin_dir_path(__FILE__));
+    define('PTC_PLUGIN_URL', plugin_dir_url(__FILE__));
 
-require_once PTC_PLUGIN_PATH.'/deprecated.php';
+    require_once __DIR__.'/deprecated.php';
 
-new Admin();
+    new ProtectTheChildren(
+        plugin_dir_path(__FILE__),
+        plugin_dir_url(__FILE__)
+    );
+});
 
 /**
  * On front-end page load, check the post's parent ID
@@ -65,7 +71,7 @@ add_action('template_redirect', function () {
     // Check the cookie (hashed password)
     require_once ABSPATH.WPINC.'/class-phpass.php';
 
-    $hasher = new PasswordHash(8, true);
+    $hasher = new \PasswordHash(8, true);
     $hash = wp_unslash($_COOKIE['wp-postpass_'.COOKIEHASH]);
     $required = ! $hasher->CheckPassword($parent_password, $hash);
 
@@ -78,26 +84,3 @@ add_action('template_redirect', function () {
         return true;
     });
 });
-
-/**
- * Upgrade meta key for older versions
- */
-function PTC_update_db_check()
-{
-    if (get_option('PTC_plugin_version', '') != PROTECT_THE_CHILDREN_PLUGIN_VERSION) {
-        $password_pages = get_pages(['meta_key' => '_protect_children', 'meta_value' => 'on']);
-
-        foreach ($password_pages as $page) {
-            update_post_meta($page->ID, 'protect_children', '1');
-            delete_post_meta($page->ID, '_protect_children');
-        }
-    }
-
-    update_option('PTC_plugin_version', PROTECT_THE_CHILDREN_PLUGIN_VERSION);
-}
-add_action('plugins_loaded', 'PTC_update_db_check');
-
-/**
- * Set version number and check for updates if necessary, on activation
- */
-register_activation_hook(__FILE__, 'PTC_update_db_check');
